Merged PR 18304: Add ALLanguage script
Adds a script that pulls the latest active dependency for the project. This avoids the need to have the "references" folder with binary dependencies.

This is a prerequisite if we want to compile the app against multiple versions. It's what will allow seamless transition between AL language versions in the future.